% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberpunk.R
\name{scale_color_cpunk}
\alias{scale_color_cpunk}
\alias{scale_colour_cpunk}
\alias{scale_fill_cpunk}
\title{scale_color_cpunk}
\usage{
scale_color_cpunk(n, type = "discrete", reverse = FALSE, ...)

scale_colour_cpunk(n, type = "discrete", reverse = FALSE, ...)

scale_fill_cpunk(n = NULL, type = "discrete", reverse = FALSE, ...)
}
\description{
scale_color_cpunk

scale_colour_cpunk

scale_fill_cpunk
}
\examples{

library(ggplot2)
ggplot(airquality, aes(x = Day, y = Temp,
     group = as.factor(Month), color = as.factor(Month))) +
     geom_point(size = 2.5) +
     scale_color_cpunk()

ggplot(airquality, aes(x = Day, y = Temp,
     group = as.factor(Month), color = as.factor(Month))) +
     geom_point(size = 2.5) +
     scale_color_cpunk()

ggplot(mpg, aes(displ)) +
     geom_histogram(aes(fill = class), col = "black", size = 0.1) +
     scale_fill_cpunk()
}
