% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberbars.R
\name{cyber_bars}
\alias{cyber_bars}
\title{Cyberpunk-style bar chart}
\usage{
cyber_bars(
  df,
  col,
  palette = 1,
  area = FALSE,
  main = NULL,
  main.size = 20,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  bg.col = "#212946",
  grid.col = "#242d4d",
  text.col = "lightgray",
  lwd = 1.75
)
}
\arguments{
\item{x}{TODO}
}
\description{
TODO
}
\examples{
# TODO

}
\author{
\itemize{
\item{Soage González, José Carlos.}
\item{Maintainer: José Carlos Soage González. \email{jsoage@uvigo.es}}
}
}
