% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyberpunklines.R
\name{cyber_lines}
\alias{cyber_lines}
\title{Cyberpunk-style line chart}
\usage{
cyber_lines(
  x,
  df,
  col,
  palette = 1,
  glow = FALSE,
  area = FALSE,
  main = NULL,
  main.size = 20,
  sub = NULL,
  xlab = NULL,
  ylab = NULL,
  bg.col = "#212946",
  grid.col = "#242d4d",
  text.col = "lightgray",
  lwd = 1.75,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{x}{TODO}
}
\description{
TODO
}
\examples{
# TODO

}
\author{
\itemize{
\item{Soage González, José Carlos.}
\item{Maintainer: José Carlos Soage González. \email{jsoage@uvigo.es}}
}
}
