% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cpunk.R
\name{theme_cpunk}
\alias{theme_cpunk}
\title{The Cyberpunk Theme#'}
\usage{
theme_cpunk(
  text.font = "mono",
  title.font = "mono",
  legend.font = "mono",
  title.size = 18,
  text.size = 14,
  subtitle.size = 12,
  axis.title.size = 14,
  axis.text.size = 10,
  legend.title.size = 10,
  legend.text.size = 9,
  title.color = "#ffffff",
  subtitle.color = "#ffffff",
  text.color = "#ffffff",
  axis.title.color = "#ffffff",
  axis.text.color = "#ffffff",
  legend.title.color = "#ffffff",
  legend.text.color = "#ffffff",
  legend.position = "bottom",
  ticks = FALSE
)
}
\arguments{
\item{text.font}{text font, Default: mono}

\item{title.font}{title font, Default: mono}

\item{legend.font}{legend font, Default: mono}

\item{title.size}{title font size, Default: 18}

\item{text.size}{text font size, Default: 14}

\item{subtitle.size}{subtitle font size, Default: 12}

\item{axis.title.size}{axis title font size, Default: 14}

\item{axis.text.size}{axis text font size, Default: 10}

\item{legend.title.size}{legend title font size, Default: 10}

\item{legend.text.size}{legend text font size, Default: 9}

\item{title.color}{title color, Default = "#ffffff"}

\item{subtitle.color}{subtitle color, Default = "#ffffff"}

\item{text.color}{text color, Default = "#ffffff"}

\item{axis.title.color}{axis title color, Default = "#ffffff"}

\item{axis.text.color}{axis text color, Default = "#ffffff"}

\item{legend.title.color}{legend title color, Default = "#ffffff"}

\item{legend.text.color}{legend text color, Default = "#ffffff"}

\item{legend.position}{legend position, Default: "bottom"}

\item{ticks}{add axis ticks, Default: FALSE}
}
\value{
ggplot2 theme object
}
\description{
The Cyberpunk Theme#'
}
\examples{
ggplot(airquality, aes(x = Day, y = Temp, group = as.factor(Month),
        color = as.factor(Month))) + geom_point(size = 2.5) +   theme_cpunk()
}
\seealso{
\link[ggplot2:theme]{ggplot2::theme}
}
