% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cybertheme.R
\name{theme_cyber}
\alias{theme_cyber}
\title{Cyberpunk ggplot2 theme}
\usage{
theme_cyber(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{x}{TODO}
}
\description{
TODO
}
\examples{
# TODO

}
\author{
\itemize{
\item{Soage González, José Carlos.}
\item{Maintainer: José Carlos Soage González. \email{jsoage@uvigo.es}}
}
}
